+++ RELEASES.md
@@ -4,9 +4,11 @@
 
 - Add pending releases here
 - Upgrade to Go version 1.24
-- Implement ACP-226: Set expected block gas cost to 0 in Granite network upgrade, removing block gas cost requirements for block building.
-- Implement ACP-226: Add `timeMilliseconds` (Unix uint64) timestamp to block header for Granite upgrade.
-- Implement ACP-226: Add `minDelayExcess` (uint64) to block header for Granite upgrade.
+- ACP-226:
+  - Set expected block gas cost to 0 in Granite network upgrade, removing block gas cost requirements for block building.
+  - Add `timeMilliseconds` (Unix uint64) timestamp to block header for Granite upgrade.
+  - Add `minDelayExcess` (uint64) to block header for Granite upgrade.
+  - Add minimum block building delays to conform the block builder to ACP-226 requirements.
 
 ## [v0.7.9](https://github.com/ava-labs/subnet-evm/releases/tag/v0.7.9)